speed to float

+++ FirstTest.py
@@ -12,7 +12,7 @@ obd.logger.setLevel(obd.logging.DEBUG)
 def get_speed(s):
     global speed
     if not s.is_null():
-        speed = int(s.value.magnitude) #for kph
+        speed = float(s.value.magnitude) #for kph
         # speed = int(s.value.magnitude * .060934)  # for mph
